+++ plugins/Level5/plugin_level5/Archives/B123.cs
@@ -125,7 +125,7 @@ namespace plugin_level5.Archives
 
             // Write file entry hashes
             _header.fileCount = fileEntries.Count;
-            _header.fileEntriesCount = (short)fileEntries.Count;
+            _header.fileEntriesCount = fileEntries.Count;
             _header.fileEntriesOffset = (int)bw.BaseStream.Position;
 
             splittedProgress[1].ReportProgress("Write file entries", 2, 4);
