+++ plugins/Level5/plugin_level5/Archives/XFSA1.cs
@@ -101,7 +101,7 @@ namespace plugin_level5.Archives
             bw.WriteAlignment(4);
 
             // Write file entry hashes
-            _header.fileEntriesCount = (short)fileEntries.Count;
+            _header.fileEntriesCount = fileEntries.Count;
             _header.fileEntriesOffset = (int)bw.BaseStream.Position;
 
             XfsaSupport.WriteCompressedTableEntries(bw.BaseStream, fileEntries.Select(x => x.Entry));
